extract gapi function
